Implemented changes as per Darren Cook:
I think the renderTests(arr:Array) function should be changed to
renderTests(Void), and then change:
               lc["addTests"](arr);
               tests = new Array();
to read:
               lc["addTests"](tests);
               tests = new Array();

The reason is simply that this.tests gets reset so passing in anything
other than this.tests is dangerous.

While there I think renderTests() and getLocalConn() can be changed from
public to private: no other classes seem to call them.

addSuccess() and addFailure() don't seem to be called from anywhere at
all. What are they for, and do they need to be part of the public interface?